ADD IT16 Confirm before deletion

+++ app/action/builtIn/delete/delete.ts
@@ -2,6 +2,8 @@ import Action  from '../../../action/action'
 import Need    from '../../../action/need'
 import Request from '../../../action/request'
 import dao     from '../../../dao/dao'
+import tr      from '../../../locale/translate'
+import Confirm from '../confirm/confirm'
 
 @Need('object')
 export default class Delete extends Action
@@ -9,6 +11,13 @@ export default class Delete extends Action
 
 	async html(request: Request)
 	{
+		const confirm   = new Confirm()
+		const confirmed = confirm.confirmed(request)
+		if (!confirmed) {
+			return confirm.html(request, tr('Do you confirm deletion') + tr('?') + "\n" + tr('All data will be lost') + '.')
+		}
+		request = confirmed
+
 		const objects = request.objects
 		for (const object of objects) {
 			await dao.delete(object)
